% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_functions.R
\name{add_labels}
\alias{add_labels}
\title{Add labels to the DESeq2 results dataframe}
\usage{
add_labels(dat)
}
\arguments{
\item{dat}{A DESeq2 results dataframe.}
}
\value{
A dataframe with labels.
}
\description{
Add labels to the DESeq2 results dataframe
}
