% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPCR_functions.R
\name{calculate_copies}
\alias{calculate_copies}
\title{Generate a Standard Curve}
\usage{
calculate_copies(data, std_list)
}
\arguments{
\item{data}{An input dataframe that has unknown samples and Cq values}

\item{std_list}{A standard curve list, direct output of \code{calculate_std_curve}, to convert Cq values to Copy Number}
}
\value{
A dataframe with a new column called \code{log10_copies} and \code{copies_rxn} (copies per reaction)
}
\description{
Take an input dataframe that has unknown samples and Cq values and uses std_list input (direct output of \code{calculate_std_curve}) to convert Cq values to Copy Number
}
