% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomics_functions.R
\name{calculate_relative_abundance}
\alias{calculate_relative_abundance}
\title{Calculate Relative Abundance from}
\usage{
calculate_relative_abundance(data, taxa_rank = "f")
}
\arguments{
\item{data}{An input dataframe containing columns titles by taxon (allowed values are: Kingdom, Phylum, Class, Order, Family, Genus, Species). The dataframe must also contain columns for each sample suffixed with "_bases" that contain coverage metrics to calculate relative abundance with.}

\item{taxa_rank}{A string indicating the taxonomic level to calculated. Allowed values are "k", "p", "c", "o", "f", "g", "s", for the first letter of each taxonomic rank.}
}
\value{
A dataframe with each taxa and the associated relative abundance.
}
\description{
Calculate Relative Abundance from
}
