% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPCR_functions.R
\name{calculate_std_curve}
\alias{calculate_std_curve}
\title{Generate a Standard Curve}
\usage{
calculate_std_curve(data)
}
\arguments{
\item{data}{An input dataframe containing at least a column called "Cq" and a column with copies called "Copies"}
}
\value{
A list containing key parameters of the standard curve and a plot of the curve.
}
\description{
Generate a Standard Curve
}
