% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_functions.R
\name{files_to_array}
\alias{files_to_array}
\title{Load files into R}
\usage{
files_to_array(strain_files)
}
\arguments{
\item{strain_files}{Output from \code{load_files()}. A string vector of relative paths to files.}
}
\value{
A dataframe.
}
\description{
Load files into R
}
