% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_functions.R
\name{format_df}
\alias{format_df}
\title{Format dataframes to be joined}
\usage{
format_df(df, file)
}
\arguments{
\item{df}{A dataframe output from \code{files_to_array()}.}

\item{file}{An input file?}
}
\value{
A dataframe.
}
\description{
Format dataframes to be joined
}
