% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_functions.R
\name{load_files}
\alias{load_files}
\title{Generate a list of files to load}
\usage{
load_files(directory, files)
}
\arguments{
\item{directory}{A directory containing all files to load.}

\item{files}{A string vector of file names to load.}
}
\value{
A list of files to load.
}
\description{
Generate a list of files to load
}
