% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_functions.R
\name{run_deseq2}
\alias{run_deseq2}
\title{Run DESeq2 and Export to and RDS File or Load DESeqDataSet Object from an Existing RDS File}
\usage{
run_deseq2(rdsName, directory)
}
\arguments{
\item{rdsName}{The desired name of your RDS or existing RDS file.}

\item{directory}{The directory to write or load your RDS file.}
}
\value{
A dataframe.
}
\description{
Run DESeq2 and Export to and RDS File or Load DESeqDataSet Object from an Existing RDS File
}
