% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomics_functions.R
\name{split_taxonomy}
\alias{split_taxonomy}
\title{Split the SQM Output Taxonomy Column into Individual Columns}
\usage{
split_taxonomy(
  data,
  taxa_list = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  sep = ";\\\\w_",
  taxa_col = "Taxon"
)
}
\arguments{
\item{data}{An input dataframe containing a columne called \code{Taxon} for which the function will split.}

\item{taxa_list}{A string vector of taxa names that will be used to rename columns.}
}
\value{
The input dataframe with the taxon column separated into the values of the \code{taxa_list}.
}
\description{
Split the SQM Output Taxonomy Column into Individual Columns
}
