% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_functions.R
\name{tidy_plate384}
\alias{tidy_plate384}
\title{Turn a 384-Well Plate Into a Tidy Dataframe}
\usage{
tidy_plate384(data, plate_id = "P1")
}
\arguments{
\item{data}{An input dataframe. Columns are numbers (1-25). The first column are the plate row names (A-P).}

\item{plate_id}{A unique identifier for a plate.}
}
\value{
A tidy dataframe with a unique plate and quadrant identifier. Can be used to join metadata onto.
}
\description{
Turns a 384-well plate where replicates are in quadrants (e.g. A1, A2, B1, B2) into a tidy dataframe where each quadrant has a unique ID
}
