% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_functions.R
\name{tidy_plate96}
\alias{tidy_plate96}
\title{Turn a 96-Well Plate Into a Tidy Dataframe}
\usage{
tidy_plate96(data, plate_id = "P1")
}
\arguments{
\item{data}{An input dataframe. Columns are numbers (1-24). The first column are the plate row names (A-P).}

\item{plate_id}{A unique identifier for a plate.}
}
\value{
A tidy dataframe with a unique plate and quadrant identifier. Can be used to join metadata onto.
}
\description{
Turn a 96-Well Plate Into a Tidy Dataframe
}
